Pipfile: fix version of dependencies
First, set the exact version of dependencies due to previous
errors with conflicts between various packages and their dependencies.

Second, the Pipfile.lock is updated because of hash mismatch. It is
strange that even though the version is the same, corresponing hash was
changed, but there is no Python requirement saying that it cannot be
done. In this case the LOCK will always help and with exact versions
set it should not destroy venv.