Add undocumented host attributes
Host object in TopologyDefinition has attributes block_internet and extra, which are
missing in the documentation. I don't know exactly what their purpose is (persumably
blocking internet access and defining extra values). If someone knows the function of these,
the description will be changed. If not, at least the documentation will mention they exist.

---v2
Attributes extra has no effect, it just allows writing extra info to the host,
and block_internet is an ancient relic that was never used and has no purpose,
so it will not be documented.

---v3
Fixed a indentation error.